+++ core/src/context.hpp
@@ -9,7 +9,7 @@
 #pragma once
 #include <list>
 #include <vector>
-#include <keyman/keyman_core_api.h>
+#include "keyman_core.h"
 
 // Forward declarations
 class json;
