+++ core/src/keyboard.hpp
@@ -11,7 +11,7 @@
 #include <string>
 #include <vector>
 
-#include <keyman/keyman_core_api.h>
+#include "keyman_core.h"
 
 #include "option.hpp"
 #include "path.hpp"
