+++ core/src/keyman_core.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <keyman/keyman_core_api.h>
+#include <keyman/keyman_core_api_context.h>
+#include <keyman/keyman_core_api_debug.h>
