+++ core/src/km_core_context_api.cpp
@@ -12,7 +12,7 @@
 #include <algorithm>
 #include <vector>
 
-#include <keyman/keyman_core_api.h>
+#include "keyman_core.h"
 
 #include "context.hpp"
 #include "jsonpp.hpp"
