+++ core/src/km_core_debug_api.cpp
@@ -8,7 +8,7 @@
 #include <algorithm>
 #include <sstream>
 
-#include <keyman/keyman_core_api.h>
+#include "keyman_core.h"
 
 #include "processor.hpp"
 #include "state.hpp"
