+++ core/src/km_core_keyboard_api.cpp
@@ -11,7 +11,8 @@
 #include <cassert>
 #include <vector>
 
-#include <keyman/keyman_core_api.h>
+#include "keyman_core.h"
+
 #include "keyboard.hpp"
 #include "processor.hpp"
 #include "kmx/kmx_processor.hpp"
