+++ core/src/km_core_state_context_set_if_needed.cpp
@@ -9,7 +9,7 @@
 */
 #include <cassert>
 
-#include <keyman/keyman_core_api.h>
+#include "keyman_core.h"
 
 #include "processor.hpp"
 #include "state.hpp"
