+++ core/src/processor.hpp
@@ -11,7 +11,7 @@
 #include <string>
 #include <unordered_map>
 
-#include <keyman/keyman_core_api.h>
+#include "keyman_core.h"
 
 #include "keyboard.hpp"
